+++ Mods ChoGGi/Challenge Show Scores/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2569673451",
 	"pops_any_uuid", "cb358485-c9bc-4d8a-84e0-2f402fd1c085",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -15,7 +15,9 @@ return PlaceObj("ModDef", {
 	"has_options", true,
 	"TagInterface", true,
 	"description", [[
-Show the completed score for challenges.
+Add more info to the challenge screen:
+Completed score
+Hours for completion instead of rounding to Sols.
 
 
 Requested by... somebody on reddit?