+++ Mods ChoGGi/Fix RC Commander Drone Freeze/metadata.lua
@@ -1,20 +1,21 @@
 return PlaceObj("ModDef", {
 	"title", "Fix RC Commander Drone Freeze",
+	"id", "ChoGGi_FixRCCommanderDroneFreeze",
+	"steam_id", "1719919254",
+	"pops_any_uuid", "a3338c37-0ad4-4beb-a102-f6b095f59a8f",
+	"lua_revision", 1007000, -- Picard
 	"version", 2,
 	"version_major", 0,
 	"version_minor", 2,
 	"image", "Preview.png",
-	"id", "ChoGGi_FixRCCommanderDroneFreeze",
-	"steam_id", "1719919254",
-	"pops_any_uuid", "a3338c37-0ad4-4beb-a102-f6b095f59a8f",
 	"author", "ChoGGi",
-	"lua_revision", 1007000, -- Picard
 	"code", {
 		"Code/Script.lua",
 	},
 	"has_options", true,
 	"TagOther", true,
-	"description", [[Fix for this bug:
+	"description", [[
+Fix for this bug:
 https://forum.paradoxplaza.com/forum/index.php?threads/surviving-mars-game-freezes-when-deploying-drones-from-rc-commander-after-one-was-destroyed.1168779/
 
 The mod can be removed afterwards.