+++ Mods ChoGGi/More Expedition RCs/metadata.lua
@@ -1,0 +1,25 @@
+return PlaceObj("ModDef", {
+	"title", "More Expedition RCs",
+	"id", "ChoGGi_MoreExpeditionRCs",
+	"steam_id", "2696123604",
+	"pops_any_uuid", "2145324d-deeb-42a3-a230-b061a35e7e74",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Relieves the limit on what counts as an RC Commander/Transport for expeditions.
+Only useful if you have Space Race DLC.
+
+
+Requested by someone awhile ago, not sure why I removed the mod?
+I suppose it's re-requested by SoftwareSimian :)
+]],
+})