+++ Mods ChoGGi/Original Amounts/metadata.lua
@@ -1,20 +1,21 @@
 return PlaceObj("ModDef", {
 	"title", "Original Amounts",
-	"version", 2,
-	"version_major", 0,
-	"version_minor", 2,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_OriginalAmounts",
 	"steam_id", "1528548594",
 	"pops_any_uuid", "01d0ffc3-d52b-425c-9466-960945afb7d6",
-	"author", "ChoGGi",
 	"lua_revision", 1007000, -- Picard
+	"version", 2,
+	"version_major", 0,
+	"version_minor", 2,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
-	"description", [[Buildings use their original build/maintence costs.
+	"description", [[
+Buildings use their original build/maintenance costs.
 Also resets supply rocket fuel cost to original.
 
-Requested by SillySMS.]],
+Requested by SillySMS.
+]],
 })