+++ Mods ChoGGi/Show Infobar Planetary VIew/metadata.lua
@@ -1,21 +1,22 @@
 return PlaceObj("ModDef", {
 	"title", "Show Infobar Planetary View",
-	"version", 2,
-	"version_major", 0,
-	"version_minor", 2,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_Show InfobarPlanetaryView",
 	"steam_id", "1743795770",
 	"pops_any_uuid", "7ab060e4-955a-4eb2-a3d2-86aaa06d41bd",
-	"author", "ChoGGi",
 	"lua_revision", 1007000, -- Picard
+	"version", 2,
+	"version_major", 0,
+	"version_minor", 2,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
-	"description", [[Shows Infobar when planetary view is opened (pushes terraform bar down).
-
-
-Requested by PaganoCristo.]],
 	"TagInterface", true,
+	"description", [[
+Shows Infobar when planetary view is opened (pushes terraform bar down).
+
+
+Requested by PaganoCristo.
+]],
 })